Week 6/Integrate LangFuse, Guardrails & Evaluate Your Agents/Multi-Agent Support System using LangGraph/requirements.txt
langchain>=0.1.20
langgraph
langchain-community
langchain-aws
chromadb
tavily-python
boto3
python-dotenv
langfuse
